Relocated Query classes
