+++ src/main/java/io/github/incplusplus/peerprocessing/client/Client.java
@@ -2,6 +2,8 @@ package io.github.incplusplus.peerprocessing.client;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import io.github.incplusplus.peerprocessing.common.*;
+import io.github.incplusplus.peerprocessing.query.AlgebraicQuery;
+import io.github.incplusplus.peerprocessing.query.Query;
 
 import java.io.BufferedReader;
 import java.io.IOException;
