+++ src/main/java/io/github/incplusplus/peerprocessing/query/MatrixQuery.java
@@ -1,0 +1,8 @@
+package io.github.incplusplus.peerprocessing.query;
+
+public class MatrixQuery extends Query {
+	@Override
+	public void complete() {
+	
+	}
+}
