+++ src/main/java/io/github/incplusplus/peerprocessing/query/Query.java
@@ -1,4 +1,4 @@
-package io.github.incplusplus.peerprocessing.common;
+package io.github.incplusplus.peerprocessing.query;
 
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
