+++ src/main/java/io/github/incplusplus/peerprocessing/slave/Slave.java
@@ -3,6 +3,8 @@ package io.github.incplusplus.peerprocessing.slave;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.udojava.evalex.Expression;
 import io.github.incplusplus.peerprocessing.common.*;
+import io.github.incplusplus.peerprocessing.query.AlgebraicQuery;
+import io.github.incplusplus.peerprocessing.query.Query;
 
 import java.io.BufferedReader;
 import java.io.IOException;
